Refactor to LivewireFileManipulationCommandParser
